Set up PyQT, moved Organisation and Simulator to another file
