+++ environment.yml
@@ -11,6 +11,8 @@ dependencies:
   - mypy=0.910
   - data-science-types=0.2.23
   - pip=21.2.2
+  - pyqt==5.12.3
+  - pyqtgraph==0.11.0
   - pip:
       - pytype==2021.07.19
       - iteround==1.0.2
