+++ refactor/src/SeaLiceMgmt.py
@@ -4,18 +4,14 @@ a given body of water.
 See README.md for details.
 """
 import argparse
-from bisect import bisect_left
-import datetime as dt
 import json
 import logging
-import dill as pickle
 import sys
 
 from pathlib import Path
 
 from src.Config import Config, to_dt
-from src.Organisation import Organisation
-from typing import Optional
+from src.Simulator import Simulator
 
 
 def create_logger():
@@ -36,69 +32,6 @@ def create_logger():
 
     return logger
 
-class Simulator:
-    def __init__(self, output_dir: Path, sim_id: str, cfg: Config):
-        self.output_dir = output_dir
-        self.sim_id = sim_id
-        self.cfg = cfg
-        self.output_dump_path = self.get_simulation_path(output_dir, sim_id)
-        self.cur_day = cfg.start_date
-        self.organisation = Organisation(cfg)
-
-    @staticmethod
-    def get_simulation_path(path: Path, sim_id: str):
-        return path / f"simulation_data_{sim_id}.pickle"
-
-    @staticmethod
-    def reload_all_dump(path: Path, sim_id: str):
-        """Reload a simulator"""
-        data_file = Simulator.get_simulation_path(path, sim_id)
-        states = []
-        times = []
-        with open(data_file, "rb") as fp:
-            try:
-                sim_state = pickle.load(fp)  # type: Simulator
-                states.append(sim_state)
-                times.append(sim_state.cur_day)
-            except EOFError:
-                pass
-
-        return states, times
-
-    @staticmethod
-    def reload(path: Path, sim_id: str, timestep: dt.datetime):
-        """Reload a simulator state from a dump at a given time"""
-        states, times = Simulator.reload_all_dump(path, sim_id)
-
-        idx = bisect_left(times, timestep)
-        return states[idx]
-
-    def run_model(self, resume=False):
-        """Perform the simulation by running the model.
-
-        :param path: Path to store the results in
-        :param sim_id: Simulation name
-        :param cfg: Configuration object holding parameter information.
-        :param Organisation: the organisation to work on.
-        """
-        cfg.logger.info("running simulation, saving to %s", self.output_dir)
-
-        # create a file to store the population data from our simulation
-        if resume and not self.output_dump_path.exists():
-            cfg.logger.warning(f"{self.output_dump_path} could not be found! Creating a new log file.")
-
-        data_file = (self.output_dump_path).open(mode="wb")
-
-        while self.cur_day <= cfg.end_date:
-            cfg.logger.debug("Current date = %s", self.cur_day)
-            self.organisation.step(self.cur_day)
-            self.cur_day += dt.timedelta(days=1)
-
-            # Save the model snapshot
-            if self.cfg.save_rate and (self.cur_day - self.cfg.start_date).days % self.cfg.save_rate == 0:
-                pickle.dump(self, data_file)
-
-        data_file.close()
 
 def generate_argparse_from_config(cfg_path: str, simulation_path: str):
     parser = argparse.ArgumentParser(description="Sea lice simulation")
@@ -181,7 +114,6 @@ if __name__ == "__main__":
     cfg = Config(cfg_path, args.param_dir, logger, vars(config_args), args.save_rate)
 
     # run the simulation
-
     if args.resume:
         resume_time = to_dt(args.resume)
         sim = Simulator.reload(output_folder, simulation_id, resume_time)  # type: Simulator
