+++ refactor/tests/conftest.py
@@ -5,7 +5,7 @@ import datetime
 
 from src.Config import Config
 from src.Farm import Farm
-from src.Organisation import Organisation
+from src.Simulator import Organisation
 from src.QueueTypes import EggBatch, DamAvailabilityBatch
 from src.LicePopulation import LicePopulation, GenoDistrib
